+++ Sustainsys.Saml2/Configuration/IdentityConfigurationsElement.cs
@@ -1,0 +1,13 @@
+﻿using System.Configuration;
+
+namespace Sustainsys.Saml2.Configuration
+{
+	public class IdentityConfigurationsElement : ConfigurationElement
+	{
+		[ConfigurationProperty("", IsDefaultCollection = true)]
+		public IdentityConfigurationElementCollection IdentityConfigurationsCollection
+		{
+			get { return (IdentityConfigurationElementCollection)this[""]; }
+		}
+	}
+}
