+++ Sustainsys.Saml2/Configuration/SecurityTokenHandlersConfigurationElement.cs
@@ -1,0 +1,13 @@
+﻿using System.Configuration;
+
+namespace Sustainsys.Saml2.Configuration
+{
+	public class SecurityTokenHandlersConfigurationElement : ConfigurationElement
+	{
+		[ConfigurationProperty("", IsDefaultCollection = true)]
+		public SecurityTokenHandlerConfigurationElementCollection SecurityTokenHandlersCollection
+		{
+			get { return (SecurityTokenHandlerConfigurationElementCollection)this[""]; }
+		}
+	}
+}
