+++ Sustainsys.Saml2/Metadata/IndexedCollectionWithDefault.cs
@@ -4,12 +4,27 @@ using System.Text;
 
 namespace Sustainsys.Saml2.Metadata
 {
+	/// <summary>
+	/// An indexed entry with an optional default
+	/// </summary>
 	public interface IIndexedEntryWithDefault
 	{
-		int Index { get; }
-		bool? IsDefault { get; }
+		/// <summary>
+		/// Index of the endpoint
+		/// </summary>
+		int Index { get; set; }
+
+		/// <summary>
+		/// Is this the default endpoint?
+		/// </summary>
+		bool? IsDefault { get; set; }
 	}
 
+	/// <summary>
+	/// A collection of indexed entries with support for getting the
+	/// configured default entry
+	/// </summary>
+	/// <typeparam name="T">The type stored in the collection</typeparam>
 	public class IndexedCollectionWithDefault<T> : SortedList<int, T> 
 		where T : class, IIndexedEntryWithDefault
 	{
